[Lens] don't update searchSession on start
