+++ x-pack/plugins/lens/public/state_management/context_middleware/subscribe_to_external_context.ts
@@ -15,6 +15,9 @@ import {
 import { setState, LensGetState, LensDispatch } from '..';
 import { getResolvedDateRange } from '../../utils';
 
+/**
+ * subscribes to external changes for filters, searchSessionId, timerange and autorefresh
+ */
 export function subscribeToExternalContext(
   data: DataPublicPluginStart,
   getState: LensGetState,
