+++ x-pack/plugins/lens/public/state_management/index.ts
@@ -9,10 +9,10 @@ import { configureStore, getDefaultMiddleware, PreloadedState } from '@reduxjs/t
 import { createLogger } from 'redux-logger';
 import { useDispatch, useSelector, TypedUseSelectorHook } from 'react-redux';
 import { makeLensReducer, lensActions } from './lens_slice';
-import { timeRangeMiddleware } from './time_range_middleware';
-import { optimizingMiddleware } from './optimizing_middleware';
 import { LensState, LensStoreDeps } from './types';
 import { initMiddleware } from './init_middleware';
+import { optimizingMiddleware } from './optimizing_middleware';
+import { contextMiddleware } from './context_middleware';
 export * from './types';
 export * from './selectors';
 
@@ -49,7 +49,7 @@ export const makeConfigureStore = (
     }),
     initMiddleware(storeDeps),
     optimizingMiddleware(),
-    timeRangeMiddleware(storeDeps.lensServices.data),
+    contextMiddleware(storeDeps),
   ];
   if (process.env.NODE_ENV === 'development') {
     middleware.push(
