+++ x-pack/plugins/lens/public/state_management/init_middleware/index.ts
@@ -7,21 +7,13 @@
 
 import { Dispatch, MiddlewareAPI, PayloadAction } from '@reduxjs/toolkit';
 import { LensStoreDeps } from '..';
-import { loadInitial as loadInitialAction, navigateAway } from '..';
+import { loadInitial as loadInitialAction } from '..';
 import { loadInitial } from './load_initial';
-import { subscribeToExternalContext } from './subscribe_to_external_context';
 
 export const initMiddleware = (storeDeps: LensStoreDeps) => (store: MiddlewareAPI) => {
-  const unsubscribeFromExternalContext = subscribeToExternalContext(
-    storeDeps.lensServices.data,
-    store.getState,
-    store.dispatch
-  );
   return (next: Dispatch) => (action: PayloadAction) => {
     if (loadInitialAction.match(action)) {
       return loadInitial(store, storeDeps, action.payload);
-    } else if (navigateAway.match(action)) {
-      return unsubscribeFromExternalContext();
     }
     next(action);
   };
