+++ x-pack/plugins/lens/public/state_management/init_middleware/load_initial.ts
@@ -99,8 +99,6 @@ export function loadInitial(
   const { resolvedDateRange, searchSessionId, isLinkedToOriginatingApp, ...emptyState } =
     getPreloadedState(storeDeps);
   const { attributeService, notifications, data, dashboardFeatureFlag } = lensServices;
-  const currentSessionId = data.search.session.getSessionId();
-
   const { lens } = store.getState();
   if (
     !initialInput ||
@@ -115,7 +113,7 @@ export function loadInitial(
           initEmpty({
             newState: {
               ...emptyState,
-              searchSessionId: currentSessionId || data.search.session.start(),
+              searchSessionId: data.search.session.getSessionId() || data.search.session.start(),
               datasourceStates: Object.entries(result).reduce(
                 (state, [datasourceId, datasourceState]) => ({
                   ...state,
@@ -178,6 +176,7 @@ export function loadInitial(
             }
           )
             .then((result) => {
+              const currentSessionId = data.search.session.getSessionId();
               store.dispatch(
                 setState({
                   isSaveable: true,
